requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==10.1.0
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6